Split up large literals
Fixes issue #256.

Having long array or object literals containing the code is problematic.
It makes it difficult to add auxiliary functions or data close to the
function inside the list or map where it is needed.
It also causes extra indentation, thus wasting space.

Now the functions, environments and symbols are no longer defined using such
a literal, but instead using a function which receives all the relevant
data.  Since each function call is independent from the others, anything can
go in between.  And since the handler is a direct argument to this, it is
indented like a top-level function.  For symbols, switching from literals to
positional arguments allowed each symbol definition to be expressed in a
single line, greatly improving the readability and compactness of the file.