Keep track of used fonts
The more symbols we have, the more fonts we ship.  And loading fonts can
take considerable time, so loading all possible fonts "just in case" is not
an option.  Therefore we keep track of the fonts actually used in a given
box, so that tools like WebFontLoader can be used to specifically load
exactly those fonts that need loading.