Add --wait option to screenshotter
Sometimes a page might not be ready when the screenshot is taken, due to
fonts still being loaded or some such.  In --verify mode this is taken care
of by repeated attempts.  But when creating a screenshot for the first time,
it might be desirable to wait a given number of seconds for things to settle
down.  So this commit introduces a --wait option to the screenshotter.