Fix positioning in limited quirks mode
This fixes #601 by adding a zero-width space to every span that contains
child nodes.  The symbol avoids the “blocks ignore line-height quirk”.

All screenshot tests render the same before and after this commit, and they
also render the same if the doctype is set to HTML 4.01 transitional in
order to trigger limited quirks mode.