Fix \boxed
A temporary patch for \boxed. When the box contains a large font, \boxed
now will pass a large value of maxFontSize to makeVList. So the box will
always be located correctly. On the other hand, the whole line may be
pushed downward on the screen.

If the all fonts inside the box are <= 1.0 em, then \boxed acts just as
in PR #670, a method which does not displace the line.

To resolve all the problems, we first need to resolve issue #746.