Fix \boxed, add \overbracket
This PR intoduces buildCommon.makeImageVList, an alternate to makeVList,
for use when one element is on the baseline and the other element is an
image. Unlike makeVList, the new function works even if image height >
the text element's font size.

The PR then uses makeImageList to fix \boxed and also adds two new
functions: \overbracket and \underbracket.